simplify install
update enc & r10k setup scripts to run from anywhere

combine reset_local & reset.sh

clone pupperware to PDIR

support for firsttime run when pupperware dir doesn't exist

dokr_compose seems to work, add to enc setup

test dokr_compose in enc setup

docker-compose fails under curl|bash

set default QS_REPO for reset

remove commeted code

+++ server/tools/reset.sh
@@ -1,24 +1,70 @@
 #!/bin/bash
 
-QS_URL=https://raw.githubusercontent.com/andylytical/quickstart/master/quickstart.sh
-export QS_REPO=https://github.com/ncsa/docker_puppet
+# SET THESE AS APPROPRIATE FOR TESTING
+# export QS_REPO=https://github.com/ncsa/docker_puppet
+# export QS_GIT_BRANCH=...
+# export ENC_GIT_BRANCH=...
+# export R10K_GIT_BRANCH=...
 
 set -x
 
-DIR="${PUPPERWARE:-$HOME/pupperware}"
-PDIR="$(readlink -e "$DIR")"
-[[ -d "$PDIR" ]] || {
-    echo "Can't find pupperware dir" >&2
-    exit 1
+TS=$(date +%s)
+QUICKSTART=https://raw.githubusercontent.com/andylytical/quickstart/master/quickstart.sh
+
+PDIR="${PUPPERWARE:-$HOME/pupperware}"
+echo "Pupperware install dir: '$PDIR'"
+COMPOSE_FILE="$PDIR/docker-compose.yml"
+
+
+# stop if QS_REPO is not set
+[[ -z "QS_REPO" ]] && QS_REPO=https://github.com/ncsa/docker_puppet
+export QS_REPO
+echo "QS_REPO=$QS_REPO"
+
+
+dokr_compose() {
+  # workaround https://github.com/docker/compose/issues/6310
+  # which says it's closed but still can't use --project-directory
+  pushd "$PDIR" &>/dev/null
+  docker-compose "$@"
+  popd &>/dev/null
 }
 
-cd "$PDIR" \
-&& docker-compose stop \
-&& docker system prune -af --volumes \
-&& cd "$PDIR"/.. \
-&& rm -rf "$PDIR" \
-&& git clone https://github.com/puppetlabs/pupperware \
-&& cd pupperware \
-&& curl $QS_URL | bash \
-&& cd "$PDIR" \
-&& docker-compose up -d
+all_services_up() {
+  local _svc_count=$(dokr_compose ps --services | wc -l)
+  local _ok_count=$(dokr_compose ps -a | tail -n+3 | grep -F 'Up (healthy)'| wc -l)
+  [[ $_ok_count -eq $_svc_count ]]
+}
+
+
+rm_existing_pupperware() {
+  [[ -d "$PDIR" ]] || return 0
+  dokr_compose stop \
+  && docker system prune -af --volumes \
+  && mv "$PDIR" "${PDIR}.$TS"
+}
+
+
+pushd ~ \
+&& rm_existing_pupperware \
+&& git clone https://github.com/puppetlabs/pupperware "$PDIR" \
+&& curl $QUICKSTART | bash \
+&& dokr_compose up -d
+popd
+
+echo "Waiting for services to start ..."
+for i in $(seq 2); do
+  sleep 30
+  all_services_up && break
+done
+all_services_up || { 
+  echo 'SERVICES NOT STARTED' >&2
+  exit 1
+}
+
+# Continue setup steps
+export COMPOSE_INTERACTIVE_NO_CLI=1
+"$PDIR"/server/enc/setup.sh \
+&& "$PDIR"/server/r10k/setup.sh \
+&& "$PDIR"/server/extras/setup.sh
+unset COMPOSE_INTERACTIVE_NO_CLI
